Refactoring. Getting rid of warnings from clang.

+++ point.h
@@ -1,10 +1,12 @@
 #ifndef POINT_H
 #define POINT_H
 
+#include <iostream>
+
 struct Point
 {
-  int row;
-  int column;
+  uint8_t row;
+  uint8_t column;
 
   Point operator+(const Point& p);
 };
